@@ -2,56 +2,81 @@
 requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
-[tool.setuptools]
-py-modules = [
-    "community_research_mcp",
-    "streaming_capabilities",
-    "streaming_search",
-    "enhanced_mcp_utilities",
-]
-packages = ["api", "core", "models", "tools", "utils"]
-
 [project]
 name = "community-research-mcp"
-version = "0.1.0"
-description = "MCP server for community-driven developer research"
+version = "1.0.0"
+description = "MCP server that searches developer communities for real-world solutions"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 license = "MIT"
-authors = [
-    {name = "DocHatty"}
+authors = [{name = "DocHatty"}]
+keywords = [
+    "mcp",
+    "search",
+    "stackoverflow",
+    "github",
+    "community",
+    "developer-tools",
 ]
-keywords = ["mcp", "search", "research", "stackoverflow", "github", "community"]
 classifiers = [
-    "Development Status :: 3 - Alpha",
+    "Development Status :: 4 - Beta",
     "Intended Audience :: Developers",
+    "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Topic :: Software Development :: Libraries",
 ]
 
 dependencies = [
     "mcp>=1.2.0",
     "fastmcp>=0.1.0",
-    "httpx>=0.28.1",
-    "pydantic>=2.10.4",
+    "httpx>=0.28.0",
+    "pydantic>=2.0.0",
     "python-dotenv>=1.0.0",
-    "beautifulsoup4>=4.13.3",
-    "google-genai>=1.19.0",
-    "openai>=1.55.2",
+    "beautifulsoup4>=4.12.0",
 ]
 
 [project.optional-dependencies]
 reddit = [
-    "praw>=7.8.1",
     "redditwarp>=1.3.0",
-    "dnspython>=2.7.0",
+]
+dev = [
+    "pytest>=7.0.0",
+    "pytest-asyncio>=0.21.0",
+    "black>=23.0.0",
+    "isort>=5.12.0",
+    "flake8>=6.0.0",
+    "mypy>=1.0.0",
 ]
 
 [project.urls]
 Homepage = "https://github.com/DocHatty/community-research-mcp"
 Repository = "https://github.com/DocHatty/community-research-mcp"
 Issues = "https://github.com/DocHatty/community-research-mcp/issues"
+
+[tool.setuptools]
+py-modules = ["community_research_mcp"]
+packages = ["api", "core", "models", "utils"]
+
+[tool.setuptools.package-data]
+"*" = ["*.json"]
+
+[tool.black]
+line-length = 88
+target-version = ["py310", "py311", "py312"]
+
+[tool.isort]
+profile = "black"
+line_length = 88
+
+[tool.mypy]
+python_version = "3.10"
+warn_return_any = true
+warn_unused_ignores = true
+ignore_missing_imports = true
+
+[tool.pytest.ini_options]
+asyncio_mode = "auto"
+testpaths = ["tests"]
